+++ pos_multi_session_restaurant/static/src/js/pos_multi_session_restaurant.js
@@ -96,13 +96,14 @@ odoo.define('pos_multi_session_restaurant', function(require){
         ms_create_order: function(options){
             var self = this;
             var order = PosModelSuper.prototype.ms_create_order.apply(this, arguments);
-            if (options.data.table_id) {
-                order.table = self.tables_by_id[options.data.table_id];
-                order.customer_count = options.data.customer_count;
+            var data = options.json;
+            if (data.table_id) {
+                order.table = self.tables_by_id[data.table_id];
+                order.customer_count = data.customer_count;
             }
             return order;
         },
-        updates_from_server: function(message, sync_all){
+        updates_from_server: function(message){
             var self = this;
             var data = message.data || {};
             var order = false;
@@ -121,7 +122,7 @@ odoo.define('pos_multi_session_restaurant', function(require){
             if ((order && old_order && old_order.uid !== order.uid) || (old_order === null)) {
                 this.set('selectedOrder',old_order);
             }
-            if (!sync_all && this.gui.screen_instances.floors && this.gui.get_current_screen() === "floors") {
+            if (this.gui.screen_instances.floors && this.gui.get_current_screen() === "floors") {
                 this.gui.screen_instances.floors.renderElement();
             }
         },
@@ -196,17 +197,6 @@ odoo.define('pos_multi_session_restaurant', function(require){
         },
     });
 
-    var MultiSessionSuper = multi_session.MultiSession;
-    multi_session.MultiSession = multi_session.MultiSession.extend({
-        sync_all: function(data) {
-            var self = this;
-            return MultiSessionSuper.prototype.sync_all.apply(this, arguments).then(function(){
-                if (self.pos.gui.screen_instances.floors && self.pos.gui.get_current_screen() === "floors") {
-                    self.pos.gui.screen_instances.floors.renderElement();
-                }
-            });
-        }
-    });
 
     floors.FloorScreenWidget.include({
         init: function(parent, options) {